Refactored project packages and files
* Refactored package names:
	- stepdefinition instead of step-definition
	- pages instead of pom
* Refactored file names of all Stepdefinitions classes
* Removed main package
