+++ .gitignore
@@ -4,10 +4,7 @@ target/
 !**/src/test/**/target/
 
 ### IntelliJ IDEA ###
-.idea/modules.xml
-.idea/jarRepositories.xml
-.idea/compiler.xml
-.idea/libraries/
+.idea
 *.iws
 *.iml
 *.ipr
@@ -33,6 +30,7 @@ target/
 /nbdist/
 /.nb-gradle/
 build/
+out/
 !**/src/main/**/build/
 !**/src/test/**/build/
