--- src/main/java/teststore/chris/App.java
@@ -1,13 +1,0 @@
-package teststore.chris;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
