+++ src/test/java/teststore/chris/pages/ArtPage.java
@@ -1,0 +1,6 @@
+package teststore.chris.pages;
+
+public class ArtPage {
+
+}
+
