+++ src/test/java/teststore/chris/pages/CartPage.java
@@ -1,4 +1,4 @@
-package teststore.chris.pom;
+package teststore.chris.pages;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
