+++ src/test/java/teststore/chris/pages/HomePage.java
@@ -1,9 +1,7 @@
-package teststore.chris.pom;
+package teststore.chris.pages;
 
-import io.cucumber.java.ro.Si;
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
 import org.openqa.selenium.support.ui.WebDriverWait;
