--- src/test/java/teststore/chris/pom/ArtPage.java
@@ -1,6 +1,0 @@
-package teststore.chris.pom;
-
-public class ArtPage {
-
-}
-
