+++ src/test/resources/logs/logFile.log
@@ -1,0 +1,3 @@
+[INFO] 2023-10-11 12:09:33.065 [main] ClothesStepdefs - Step failed - Wrong page
+[WARN] 2023-10-11 12:09:33.067 [main] ClothesStepdefs - Expected page title: Clothesst
+[WARN] 2023-10-11 12:09:33.068 [main] ClothesStepdefs - Actual page title: Clothes
